refactor(solana): adds better util file naming
 - check if rpc connections exists before creation
